11/8

+++ 레이디버그_넷겜플/레이디버그_넷겜플/소스.cpp
@@ -116,8 +116,8 @@ typedef struct Move {
 static struct BUG
 {
 	float x;
-	int y;
-	int y_move;
+	float y;
+	float y_move;
 	float x_move;
 	int w;
 	int h;
